django/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asgiref=3.4.1=pyhd3eb1b0_0
blas=1.0=mkl
ca-certificates=2021.10.26=hecd8cb5_2
certifi=2021.10.8=py39hecd8cb5_0
django=3.2.5=pyhd3eb1b0_0
django-cors-headers=3.10.0=pypi_0
djangorestframework=3.12.4=pyhd8ed1ab_0
intel-openmp=2021.4.0=hecd8cb5_3538
krb5=1.19.2=hcd88c3b_0
libcxx=12.0.0=h2f01273_0
libedit=3.1.20210910=hca72f7f_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libpq=12.2=h1b4eb34_1
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
ncurses=6.3=hca72f7f_2
numpy=1.21.2=py39h4b4dc7a_0
numpy-base=1.21.2=py39he0bd621_0
openssl=1.1.1l=h9ed2024_0
pip=21.2.4=py39hecd8cb5_0
psycopg2=2.8.6=py39hbcfaee0_1
python=3.9.7=h88f2d9e_1
python_abi=3.9=2_cp39
pytz=2021.3=pyhd3eb1b0_0
readline=8.1=h9ed2024_0
scipy=1.7.1=py39h88652d9_2
setuptools=58.0.4=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hce871da_0
sqlparse=0.4.1=py_0
tk=8.6.11=h7bc2e8c_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
